Move process transaction to tasks
